add to cart functionality done

+++ database/migrations/2020_07_15_102626_create_carts_table.php
@@ -22,7 +22,7 @@ class CreateCartsTable extends Migration
             $table->float('price')->nullable();
             $table->enum('status',['new','progress','delivered','cancel'])->default('new');
             $table->integer('quantity')->nullable();
-            $table->float('amount')->nullable();
+            $table->unsignedBigInteger('amount')->nullable();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('CASCADE');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE');
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('SET NULL');
